refactor(grocery): move GroceryDetail previews + test tags into impl-friendly locations
GroceryDetailSheetContent now lives in client/grocery/core/impl after
the BlocScreen pattern landed on main, so the preview blocs move there
too — and pick up the new BlocScreen contract via a Content override
that routes to GroceryDetailSheetContent. Test tags are split out into
their own files under :public so impl-robots can match against them
without depending on impl.
