+++ client/grocery/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/grocery/core/detail/GroceryDetailTestTags.kt
@@ -1,0 +1,6 @@
+package com.plusmobileapps.chefmate.grocery.core.detail
+
+object GroceryDetailTestTags {
+    const val SHEET = "grocery_detail_sheet"
+    const val AISLE_DROPDOWN = "grocery_detail_aisle_dropdown"
+}
