+++ client/grocery/core/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/grocery/core/list/GroceryListTestTags.kt
@@ -1,0 +1,5 @@
+package com.plusmobileapps.chefmate.grocery.core.list
+
+object GroceryListTestTags {
+    const val SCREEN = "grocery_list_screen"
+}
