+++ client/ui/screenshot-test/src/screenshotTest/kotlin/com/plusmobileapps/chefmate/ui/screenshot/GroceryDetailScreenshotTest.kt
@@ -12,11 +12,11 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.android.tools.screenshot.PreviewTest
 import com.plusmobileapps.chefmate.grocery.core.detail.GroceryDetailBloc
-import com.plusmobileapps.chefmate.grocery.core.detail.GroceryDetailSheetContent
-import com.plusmobileapps.chefmate.grocery.core.detail.previewGroceryDetailBlocFromRecipe
-import com.plusmobileapps.chefmate.grocery.core.detail.previewGroceryDetailBlocLoaded
-import com.plusmobileapps.chefmate.grocery.core.detail.previewGroceryDetailBlocLoading
-import com.plusmobileapps.chefmate.grocery.core.detail.previewGroceryDetailBlocPurchased
+import com.plusmobileapps.chefmate.grocery.core.impl.detail.ui.GroceryDetailSheetContent
+import com.plusmobileapps.chefmate.grocery.core.impl.detail.ui.previewGroceryDetailBlocFromRecipe
+import com.plusmobileapps.chefmate.grocery.core.impl.detail.ui.previewGroceryDetailBlocLoaded
+import com.plusmobileapps.chefmate.grocery.core.impl.detail.ui.previewGroceryDetailBlocLoading
+import com.plusmobileapps.chefmate.grocery.core.impl.detail.ui.previewGroceryDetailBlocPurchased
 import com.plusmobileapps.chefmate.text.FixedString
 import com.plusmobileapps.chefmate.ui.components.PlusResponsiveModalDialogSurface
 import com.plusmobileapps.chefmate.ui.theme.ChefMateTheme
